fix(ci): python-coverage-shards — split oversized test_artifacts.py to satisfy maintainability line limit
test_artifacts.py had grown to 1539 lines, tripping the
test_first_party_code_files_stay_under_line_limit guardrail (1500-line
cap). Move the TestDepositWorkspacePlanningArtifacts suite into a new
test_artifacts_deposit.py sibling and prune the now-unused imports from
each file. No test behavior changes; the same 69 tests run and pass.
